+++ flows/interfaces.go
@@ -52,23 +52,6 @@ const NilMsgID = MsgID(0)
 // MsgUUID is the UUID of a message
 type MsgUUID uuids.UUID
 
-// FlowType represents the different types of flows
-type FlowType string
-
-const (
-	// FlowTypeMessaging is a flow that is run over a messaging channel
-	FlowTypeMessaging FlowType = "messaging"
-
-	// FlowTypeMessagingPassive is a non-interactive messaging flow (i.e. never waits for input)
-	FlowTypeMessagingPassive FlowType = "messaging_passive"
-
-	// FlowTypeMessagingOffline is a flow which is run over an offline messaging client like Surveyor
-	FlowTypeMessagingOffline FlowType = "messaging_offline"
-
-	// FlowTypeVoice is a flow which is run over IVR
-	FlowTypeVoice FlowType = "voice"
-)
-
 // SessionStatus represents the current status of the engine session
 type SessionStatus string
 
@@ -171,20 +154,6 @@ type Node interface {
 	EnumerateDependencies(Localization, func(Action, Router, envs.Language, assets.Reference))
 	EnumerateResults(func(Action, Router, *ResultInfo))
 	EnumerateLocalizables(func(uuids.UUID, string, []string, func([]string)))
-}
-
-// FlowTypeRestricted is a part of a flow which can be restricted to certain flow types
-type FlowTypeRestricted interface {
-	AllowedFlowTypes() []FlowType
-}
-
-func (t FlowType) Allows(r FlowTypeRestricted) bool {
-	for _, allowedType := range r.AllowedFlowTypes() {
-		if t == allowedType {
-			return true
-		}
-	}
-	return false
 }
 
 // Action is an action within a flow node
